Import cleanup and making things a bit more reasonable
