+++ heroku_applink/__init__.py
@@ -5,50 +5,15 @@ SPDX-License-Identifier: BSD-3-Clause
 For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
 """
 
-from contextvars import ContextVar
-
 from .authorization import Authorization
 from .config import Config
-from .context import ClientContext
+from .context import ClientContext, get_client_context, set_client_context
 from .data_api.record import QueriedRecord, Record, RecordQueryResult
 from .data_api.reference_id import ReferenceId
 from .data_api.unit_of_work import UnitOfWork
 from .middleware import IntegrationWsgiMiddleware, IntegrationAsgiMiddleware
 from .exceptions import ClientError, UnexpectedRestApiResponsePayload
 from .connection import Connection
-
-# ContextVars for request-scoped data
-client_context: ContextVar[ClientContext] = ContextVar("client_context")
-request_id: ContextVar[str] = ContextVar("request_id")
-
-def get_client_context() -> ClientContext:
-    """
-    Call `get_client_context` to get the client context for the current incoming
-    request from Salesforce. This will be set by the `IntegrationWsgiMiddleware` or
-    `IntegrationAsgiMiddleware` in your application and can only be used in requests
-    that are routed through one of these middlewares.
-
-    ```python
-    import heroku_applink as sdk
-    from fastapi import FastAPI
-
-    app = FastAPI()
-    app.add_middleware(sdk.IntegrationAsgiMiddleware, config=sdk.Config(request_timeout=5))
-
-    @app.get("/accounts")
-    async def get_accounts():
-        context = sdk.get_client_context()
-
-        query = "SELECT Id, Name FROM Account"
-        result = await context.data_api.query(query)
-
-        return {"accounts": [record.get("Name") for record in result.records]}
-    ```
-    """
-    try:
-      return client_context.get()
-    except LookupError:
-        raise ValueError("No client context found")
 
 def get_authorization(developer_name: str, attachment_or_url: str|None=None) -> Authorization:
     """
@@ -74,12 +39,11 @@ def get_authorization(developer_name: str, attachment_or_url: str|None=None) -> 
 
 __all__ = [
     "get_client_context",
+    "set_client_context",
     "get_authorization",
     "Authorization",
     "Config",
     "Connection",
-    "client_context",
-    "request_id",
     "ClientContext",
     "QueriedRecord",
     "Record",
