+++ tests/test_asgi_api.py
@@ -16,7 +16,7 @@ async def root():
 
 @app.get("/client-context")
 async def client_context():
-    data_api = sdk.client_context.get().data_api
+    data_api = sdk.get_client_context().data_api
 
     return {"data_api_populated": data_api is not None}
